src/requirements.txt
aioredis==1.1.0
asgiref==2.3.0
astroid==1.6.5
async-timeout==2.0.1
attrs==18.1.0
autobahn==18.5.1
Automat==0.6.0
certifi==2018.4.16
channels==2.1.1
channels-redis==2.2.0
chardet==3.0.4
constantly==15.1.0
daphne==2.1.1
defusedxml==0.5.0
Django==2.0.5
django-allauth==0.36.0
django-cors-headers==2.3.0
django-environ==0.4.5
django-filter==1.1.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
hiredis==0.2.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==2.6.11
mccabe==0.6.1
msgpack==0.5.6
oauthlib==2.1.0
Pillow==5.1.0
pkg-resources==0.0.0
psycopg2==2.7.5
PyJWT==1.6.4
pylint==1.9.2
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.4
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
Twisted==18.4.0
txaio==2.10.0
urllib3==1.23
uWSGI==2.0.17
wrapt==1.10.11
zope.interface==4.5.0
